+++ src/lib.rs
@@ -1343,6 +1343,10 @@ pub struct GooseConfiguration {
     #[structopt(long, default_value = "goose.log")]
     pub log_file: String,
 
+    /// Prevent load test client from following redirect of base_url with subsequent requests
+    #[structopt(long)]
+    pub no_sticky_follow: bool,
+
     /// Enables manager mode
     #[structopt(long)]
     pub manager: bool,